worker/requirements.txt
aio-pika==9.4.0
aiormq==6.8.1
idna==3.10
multidict==6.4.3
pamqp==3.3.0
propcache==0.3.1
python-dotenv==1.0.0
typing_extensions==4.13.2
yarl==1.20.0
